@@ -15,12 +15,15 @@ using AppBokerASP.Database;
 using AppBokerASP.Database.Model;
 using AppBokerASP.Devices;
 using AppBokerASP.Devices.Heater;
+using AppBokerASP.IOBroker;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 using PainlessMesh;
-
+using SimpleSocketIoClient;
 
 namespace AppBokerASP
 {
@@ -35,17 +38,41 @@ namespace AppBokerASP
         //private static TcpListener TcpServer = new TcpListener(new IPAddress(new byte[] { 0, 0, 0, 0 }), 8801);
 
 
-
+        /*
+         * 
+         * 
+    [JsonPropertyName("id"), nj.JsonProperty("id")]
+    public uint NodeId { get; set; }
+    [System.Text.Json.Serialization.JsonConverter(typeof(JsonStringEnumConverter)), JsonPropertyName("m"), nj.JsonProperty("m")]
+    public MessageType MessageType { get; set; }
+    [System.Text.Json.Serialization.JsonConverter(typeof(JsonStringEnumConverter)), JsonPropertyName("c"), nj.JsonProperty("c")]
+    public Command Command { get; set; }
+    [JsonPropertyName("p"), nj.JsonProperty("p")/*, nj.JsonConverter(typeof(SingleOrListConverter<JsonElement>))]
+    public List<JObject> Parameters { get; set; }
+    public GeneralSmarthomeMessage(uint nodeId, MessageType messageType, Command command, params JObject[] parameters)
+    {
+        NodeId = nodeId;
+        MessageType = messageType;
+        Command = command;
+        Parameters = parameters.ToList();
+    }
+    public GeneralSmarthomeMessage()
+    {
+    }
+         * 
+         * 
+         * 
+         */
         public static void Main(string[] args)
         {
             //TimeTempMessageLE.Test();
             //var ttm = new TimeTempMessageLE(Devices.Heater.DayOfWeek.Sat, TimeSpan.FromMinutes(1985), 55.5f);
             //Console.WriteLine(ttm.GetBits());
             Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
-            var e =Encoding.GetEncodings();
+
+            Dostuff();
             MeshManager = new SmarthomeMeshManager(8801);
             DeviceManager = new DeviceManager();
-            Dostuff();
 
             int asd = 123;
             var bytes = BitConverter.GetBytes(asd);
@@ -56,9 +83,9 @@ namespace AppBokerASP
             else
             {
 #if DEBUG
-                CreateWebHostBuilder(args).UseUrls("http://[::1]:5056", "http://192.168.49.28:5056").Build().Run();
+                CreateWebHostBuilder(args).UseUrls("http://[::1]:5056", "http://0.0.0.0:5056").Build().Run();
 #else
-                CreateWebHostBuilder(args).UseUrls("http://[::1]:5055", "http://192.168.49.71:5055").Build().Run();
+                CreateWebHostBuilder(args).UseUrls("http://[::1]:5055", "http://0.0.0.0:5055").Build().Run();
 #endif
             }
         }
@@ -68,7 +95,7 @@ namespace AppBokerASP
         private static async void Dostuff()
         {
 
-
+        
             return;
             var wc = WebRequest.Create(@"https://192.168.49.71:8087/objects?pattern=system.adapter.zigbee.0*&prettyPrint");
 
