@@ -8,46 +8,51 @@ using System.Net.Sockets;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
+using Newtonsoft.Json.Linq;
 using PainlessMesh;
 
 namespace AppBokerASP
 {
     public class SmarthomeMeshManager
     {
-        private TimeSpan WaitBeforeWhoIAmSendAgain = new TimeSpan(0, 0, 30);
 
         public event EventHandler<GeneralSmarthomeMessage> SingleUpdateMessageReceived;
+        public event EventHandler<GeneralSmarthomeMessage> SingleOptionsMessageReceived;
         public event EventHandler<GeneralSmarthomeMessage> SingleGetMessageReceived;
         public event EventHandler<(Sub, List<string>)> NewConnectionEstablished;
+        public event EventHandler<uint> ConnectionLost;
 
         private static ServerSocket serverSocket = new ServerSocket();
 
+        private readonly TimeSpan WaitBeforeWhoIAmSendAgain;
         private readonly List<uint> knownNodeIds;
-        private readonly ConcurrentDictionary<uint, DateTime> WhoIAmSendTime;
+        private readonly ConcurrentDictionary<uint, (DateTime time, int count)> WhoIAmSendTime;
 
         private readonly uint nodeID = 1;
+        private Dictionary<uint, Queue<GeneralSmarthomeMessage>> queuedMessages;
 
+        public List<Timer> Timer;
 
-        private Timer timeBroadcastTimer;
-        private Timer whoAmITask;
-        private Timer getMeshTreeUpdateTask;
 
         public SmarthomeMeshManager(int listenPort, uint nodeId = 1)
         {
+            WaitBeforeWhoIAmSendAgain = new TimeSpan(0, 0, 30);
+            queuedMessages = new Dictionary<uint, Queue<GeneralSmarthomeMessage>>();
+            Timer = new List<Timer>();
             nodeID = nodeId;
-            WhoIAmSendTime = new ConcurrentDictionary<uint, DateTime>();
-            knownNodeIds = new List<uint> { nodeID };
+            WhoIAmSendTime = new ConcurrentDictionary<uint, (DateTime, int)>();
+            knownNodeIds = new List<uint> { nodeID, 0 };
             serverSocket.OnClientConnected += ServerSocket_OnClientConnected;
             serverSocket.Start(new IPAddress(new byte[] { 0, 0, 0, 0 }), listenPort);
-            whoAmITask = new Timer(WhoAmITask, null, TimeSpan.FromSeconds(10d), WaitBeforeWhoIAmSendAgain);
-            timeBroadcastTimer = new Timer((n) => SendBroadcast(new GeneralSmarthomeMessage(0, MessageType.Update, Command.Time, JsonSerializer.Deserialize<JsonElement>($"{{\"Date\":\"{DateTime.Now.ToString()}\"}}"))), null, TimeSpan.FromMinutes(1d), TimeSpan.FromDays(1d));
-            getMeshTreeUpdateTask = new Timer((n) => SendToBridge(new GeneralSmarthomeMessage(1, MessageType.Get, Command.Mesh, JsonSerializer.Deserialize<JsonElement>("{}"))), null, TimeSpan.FromSeconds(20d), TimeSpan.FromMinutes(1d));
+            Timer.Add(new Timer(WhoAmITask, null, TimeSpan.FromSeconds(10d), WaitBeforeWhoIAmSendAgain));
+            Timer.Add(new Timer((n) => SendBroadcast(new GeneralSmarthomeMessage(0, MessageType.Update, Command.Time, $"{{\"Date\":\"{DateTime.Now.ToString("dd.MM.yyyy HH:mm:ss")}\"}}".ToJToken())), null, TimeSpan.FromMinutes(1d), TimeSpan.FromHours(1d)));
+            Timer.Add(new Timer((n) => SendToBridge(new GeneralSmarthomeMessage(1, MessageType.Get, Command.Mesh)), null, TimeSpan.FromSeconds(20d), TimeSpan.FromMinutes(1d)));
         }
 
         private void ServerSocket_OnClientConnected(object sender, BaseClient baseClient)
         {
             //baseClient.Send(PackageType.BRIDGE, "GetMeshTree", nodeID);
-            SendToBridge(new GeneralSmarthomeMessage(1, MessageType.Get, Command.Mesh, JsonSerializer.Deserialize<JsonElement>("{}")));
+            SendToBridge(new GeneralSmarthomeMessage(1, MessageType.Get, Command.Mesh));
             baseClient.ReceivedData += SocketClientDataReceived;
             //baseClient.Start();
         }
@@ -71,17 +76,42 @@ namespace AppBokerASP
 
             if (!knownNodeIds.Contains(e.NodeId))
             {
-                if (!WhoIAmSendTime.TryGetValue(e.NodeId, out var dt) || dt.Add(WaitBeforeWhoIAmSendAgain) > DateTime.Now)
+                if (!WhoIAmSendTime.TryGetValue(e.NodeId, out var dt) || dt.time.Add(WaitBeforeWhoIAmSendAgain) > DateTime.Now)
                 {
                     //SendSingle(e.NodeId, new GeneralSmarthomeMessage(0, MessageType.Get, Command.WhoIAm));
                     if (dt == default)
-                        while (!WhoIAmSendTime.TryAdd(e.NodeId, DateTime.Now)) { }
+                        while (!WhoIAmSendTime.TryAdd(e.NodeId, (DateTime.Now, 0))) { }
                     else
-                        WhoIAmSendTime[e.NodeId] = DateTime.Now;
+                        WhoIAmSendTime[e.NodeId] = (DateTime.Now, 0);
                 }
+                if (!queuedMessages.TryGetValue(e.NodeId, out var queue))
+                {
+                    queue = new Queue<GeneralSmarthomeMessage>();
+                    queuedMessages.Add(e.NodeId, queue);
+                }
+                queue.Enqueue(e);
                 return;
             }
 
+            if (queuedMessages.TryGetValue(e.NodeId, out var messages))
+            {
+                while (messages.TryDequeue(out var message))
+                    switch (message.MessageType)
+                    {
+                        case MessageType.Get:
+                            HandleGets(message);
+                            break;
+                        case MessageType.Update:
+                            HandleUpdates(message);
+                            break;
+                        case MessageType.Options:
+                            HandleOptions(message);
+                            break;
+                        default:
+                            break;
+                    }
+                queuedMessages.Remove(e.NodeId);
+            }
 
             switch (e.MessageType)
             {
@@ -91,28 +121,40 @@ namespace AppBokerASP
                 case MessageType.Update:
                     HandleUpdates(e);
                     break;
+                case MessageType.Options:
+                    HandleOptions(e);
+                    break;
                 default:
                     break;
             }
         }
 
-        private void ParseSubs(Sub sub)
+        internal void UpdateTime()
+        => SendBroadcast(new GeneralSmarthomeMessage(0, MessageType.Update, Command.Time, $"{{\"Date\":\"{DateTime.Now.ToString("dd.MM.yyyy HH:mm:ss")}\"}}".ToJToken()));
+
+
+        private List<Sub> ParseSubsRecursive(Sub sub)
         {
+            var subs = new List<Sub>();
+            subs.Add(sub);
             if (!knownNodeIds.Contains(sub.NodeId))
             {
-                if (!WhoIAmSendTime.TryGetValue(sub.NodeId, out var dt) || dt.Add(WaitBeforeWhoIAmSendAgain) > DateTime.Now)
+                if (!WhoIAmSendTime.TryGetValue(sub.NodeId, out var dt) || dt.time.Add(WaitBeforeWhoIAmSendAgain) > DateTime.Now)
                 {
                     //SendSingle(sub.NodeId, new GeneralSmarthomeMessage(0, MessageType.Get, Command.WhoIAm));
                     if (dt == default)
-                        while (!WhoIAmSendTime.TryAdd(sub.NodeId, DateTime.Now)) { }
+                        while (!WhoIAmSendTime.TryAdd(sub.NodeId, (DateTime.Now, 0))) { }
                     else
-                        WhoIAmSendTime[sub.NodeId] = DateTime.Now;
+                        WhoIAmSendTime[sub.NodeId] = (DateTime.Now, 0);
                 }
             }
+
             if (sub.Subs != null && sub.Subs != default)
                 foreach (var item in sub.Subs)
-                    ParseSubs(item.Value);
+                    subs.AddRange(ParseSubsRecursive(item.Value));
+            return subs;
         }
+
 
         private void HandleUpdates(GeneralSmarthomeMessage e)
         {
@@ -123,8 +165,29 @@ namespace AppBokerASP
                 case Command.OnChangedConnections:
                 case Command.OnNewConnection:
                 case Command.Mesh:
+
                     var sub = e.Parameters[0].ToObject<Sub>();
-                    ParseSubs(sub);
+                    var subs = ParseSubsRecursive(sub).ToDictionary(x => x.NodeId, x => x);
+                    var subIds = new List<uint>();
+
+                    foreach (var item in knownNodeIds)
+                    {
+                        if (!subs.ContainsKey(item))
+                            subIds.Add(item);
+                    }
+                    foreach (var item in WhoIAmSendTime)
+                    {
+                        if (subIds.Contains(item.Key))
+                        {
+                            if (subs.ContainsKey(item.Key))
+                                subIds.Remove(item.Key);
+                        }
+                    }
+                    foreach (var id in subIds)
+                    {
+                        ConnectionLost?.Invoke(this, id);
+                        knownNodeIds.Remove(id);
+                    }
                     break;
                 default:
                     SingleUpdateMessageReceived?.Invoke(this, e);
@@ -136,32 +199,79 @@ namespace AppBokerASP
         private void HandleGets(GeneralSmarthomeMessage e)
         {
             if (e.Command == Command.Time)
-                SendSingle(e.NodeId, new GeneralSmarthomeMessage(e.NodeId, MessageType.Update, Command.Time, JsonSerializer.Deserialize<JsonElement>($"{{\"Date\":\"{DateTime.Now.ToString()}\"}}")));
+                SendBroadcast(new GeneralSmarthomeMessage(e.NodeId, MessageType.Update, Command.Time, $"{{\"Date\":\"{DateTime.Now.ToString("dd.MM.yyyy HH:mm:ss")}\"}}".ToJToken()));
             else
                 SingleGetMessageReceived?.Invoke(this, e);
         }
 
+        private void HandleOptions(GeneralSmarthomeMessage e)
+        {
+            SingleOptionsMessageReceived?.Invoke(this, e);
+        }
+
         public void SendSingle<T>(uint destination, T message)
-            => serverSocket.SendToAllClients(PackageType.SINGLE, message.ToJson(), destination);
+        {
+            try
+            {
+                serverSocket.SendToAllClients(PackageType.SINGLE, message.ToJson(), destination);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+            }
+        }
 
         public void SendBroadcast<T>(T message)
-            => serverSocket.SendToAllClients(PackageType.BROADCAST, message.ToJson());
+        {
+            try
+            {
+                serverSocket.SendToAllClients(PackageType.BROADCAST, message.ToJson());
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+            }
+        }
 
         public void SendToBridge<T>(T message)
-            => serverSocket.SendToAllClients(PackageType.BRIDGE, message.ToJson(), nodeID);
+        {
+            try
+            {
+                serverSocket.SendToAllClients(PackageType.BRIDGE, message.ToJson(), nodeID);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+            }
+        }
 
         private void WhoAmITask(object o)
         {
-
-            foreach (var item in WhoIAmSendTime)
+            var toDelete = new Dictionary<uint, (DateTime time, int count)>();
+            try
             {
-                if (item.Value.Add(WaitBeforeWhoIAmSendAgain) < DateTime.Now)
+                foreach (var item in WhoIAmSendTime)
                 {
-                    SendSingle(item.Key, new GeneralSmarthomeMessage(0, MessageType.Get, Command.WhoIAm));
-                    WhoIAmSendTime[item.Key] = DateTime.Now;
+                    if (item.Value.time.Add(WaitBeforeWhoIAmSendAgain) < DateTime.Now)
+                    {
+                        SendSingle(item.Key, new GeneralSmarthomeMessage(0, MessageType.Get, Command.WhoIAm));
+                        WhoIAmSendTime[item.Key] = (DateTime.Now, item.Value.count + 1);
+                    }
+                    if (item.Value.count > 20)
+                        toDelete.Add(item.Key, item.Value);
                 }
-            }
 
+                foreach (var item in toDelete)
+                {
+                    while (!WhoIAmSendTime.TryRemove(item.Key, out var val))
+                        Thread.Sleep(1);
+                }
+                toDelete.Clear();
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+            }
         }
 
     }
